Found the mistake.
- the problem was that I passed a pointer to an automatic variable which
was changed after i exited the isMemory command function. Fixed, now
just reruns the same line and erases all [ and ] from it. After that it
runs jumpToLastSpace so that %n in sscanf always gives the same result

- checking the argument (the number itself) in [ax+5] and ax+5 was
changed so it doesn't depend on the length of a number (%n counts only
upon +, after that we just read the command)